+++ .classpath
@@ -1,13 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="src" path="src"/>
-<<<<<<< HEAD
+	<classpathentry kind="src" path=""/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
 	<classpathentry kind="lib" path="mysql-connector-java-5.1.13.jar"/>
-=======
-	<classpathentry kind="lib" path="C:/Users/Armin/git/360Project/mysql-connector-java-5.1.13.jar"/>
-	<classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.module.container"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
->>>>>>> 52b69404b7eaf1ec9cd6b662d2b7df3d7e22dbfa
 	<classpathentry kind="output" path="bin"/>
 </classpath>
